+++ common/src/main/java/ru/civilea/common/models/CreateCityDto.kt
@@ -4,5 +4,5 @@ import java.io.Serializable
 
 data class CreateCityDto(
     val name: String = "",
-    val weatherDegree: Long = 0
+    val weatherDegree: Int = 0
 ) : Serializable
