+++ server/src/db/table/Cities.kt
@@ -6,5 +6,5 @@ import org.jetbrains.exposed.sql.Table
 object Cities : Table() {
     var id: Column<Long> = Cities.long("id").autoIncrement().primaryKey()
     var name: Column<String> = Cities.text("name")
-    var weatherDegree: Column<Long> = Cities.long("weatherdegree")
+    var weatherDegree: Column<Int> = Cities.integer("weatherdegree")
 }
